feat: create useCustomMutation hook for simple mutations
