+++ frontend/src/lib/types/global.ts
@@ -30,4 +30,12 @@ export interface SetRelayStatusPayload {
   status: boolean;
 }
 
+//TODO: consider typed query keys
+// export enum QueryKeys {
+//   Devices = 'devices',
+//   Rooms = 'rooms',
+//   Relay = 'Relay',
+//   DHTSensor = 'DHTSensor',
+// }
+
 export type addDeviceItemsGroupType = typeof newDeviceTypeSelectItems;
